Update pylint evaluation function
If we want to make sure our codebase stays clean we must avoid possible
rounding errors in the scoring system. The 0 to 10 range can still show
the perfect 10.0 score if there are enough lines of code and very few
errors. By using the count method we can be sure that 0 is the best
score and any increase is easily visible.